requirements.txt
streamlit==1.32.0
google-generativeai==0.3.2
pdfplumber==0.10.3
scikit-learn==1.3.2
numpy==1.26.0